+++ CUDACore/src/debug.jl
@@ -1,4 +1,0 @@
-# debug functionality
-
-isdebug(group, mod=CUDACore) =
-    Base.CoreLogging.current_logger_for_env(Base.CoreLogging.Debug, group, mod) !== nothing